contour: put each notification concern at the layer that owns it
Two placements were left over from before the portal backend existed.

NotificationTransport's own doc says the seam is "stated in the DOMAIN type,
not in either wire format -- each implementation builds its own arguments",
and buildFreedesktopNotifyArguments/buildPortalNotificationOptions follow it.
The two urgency tables did not: both sat in the shared NotificationRouter,
each with exactly one caller elsewhere. Splitting the bookkeeping out into
NotificationIdMap left the router as routing policy, which makes two wire
tables inside it the more anomalous. They move beside the call that carries
them, so a third backend adds its table to its own header.

makePortalIdPrefix() likewise sat in Notifier.hpp, which every platform
includes, and compiled its QUuid on Windows and macOS where nothing can call
it. It is not a decision to be shared -- it is one transport's id scheme, and
that transport's header already documents why it needs a namespace.

Their tests move to the already platform-guarded files.
