+++ src/contour/platform/Notifier.hpp
@@ -103,22 +103,6 @@ enum class NotificationBackend : uint8_t
     return NotificationBackend::FreeDesktop;
 }
 
-/// A fresh portal-notification id namespace, unique within this process and across calls.
-///
-/// Portal notification ids are chosen by the application, and two things make a namespace
-/// necessary. Between processes: the portal's ActionInvoked signal is broadcast and carries no app
-/// id, so another sandboxed application's notification "1" would be indistinguishable from ours.
-/// WITHIN one process: a transport is built per terminal session, and two panes are free to use the
-/// same OSC 99 identifier -- so a per-process prefix alone would have them collide, replacing each
-/// other's popups and each answering the other's activation. Hence unique per call, not per process.
-///
-/// The per-process half is random rather than the process id, because a sandbox is precisely where
-/// a process id stops identifying anything: Flatpak gives every instance its own PID namespace, so
-/// two sandboxed Contours both see a small, identical pid.
-///
-/// @return The prefix, of the shape `contour-<uuid>-<n>`.
-[[nodiscard]] std::string makePortalIdPrefix();
-
 /// The transport for a given backend.
 ///
 /// Separate from makeDesktopNotifier() so the dispatch is reachable without the sandbox that would
