+++ src/contour/platform/PortalNotificationTransport.cpp
@@ -6,11 +6,13 @@
     #include <contour/platform/PortalNotificationTransport.hpp>
 
     #include <QtCore/QTimer>
+    #include <QtCore/QUuid>
     #include <QtDBus/QDBusMessage>
     #include <QtDBus/QDBusPendingCallWatcher>
     #include <QtDBus/QDBusPendingReply>
 
     #include <array>
+    #include <atomic>
     #include <format>
     #include <utility>
 
@@ -53,7 +55,7 @@ namespace
 
 QVariantMap buildPortalNotificationOptions(vtbackend::DesktopNotification const& notification)
 {
-    auto const priority = NotificationRouter::toPortalPriority(notification.urgency);
+    auto const priority = toPortalPriority(notification.urgency);
 
     auto options = QVariantMap {};
 
@@ -105,6 +107,23 @@ PortalCaller qtPortalCaller()
                 onReply(reply.isError() ? CallOutcome::Failed : CallOutcome::Accepted);
             });
     };
+}
+
+std::string makePortalIdPrefix()
+{
+    // A random per-process component rather than the process id, because the process id is exactly
+    // what a sandbox takes away: Flatpak runs every instance in its own PID namespace, so two
+    // Contour instances -- the only case where a collision between processes is possible at all --
+    // both report a small, identical pid. A UUID is unique wherever a pid is, and stays unique
+    // where a pid stops being.
+    static auto const processNamespace = QUuid::createUuid().toString(QUuid::WithoutBraces).toStdString();
+
+    // A plain counter rather than anything meaningful: the only requirement is that no two
+    // transports in this process share a namespace. Atomic because sessions are created from more
+    // than one thread.
+    static auto nextNamespace = std::atomic<uint64_t> { 0 };
+
+    return std::format("contour-{}-{}", processNamespace, ++nextNamespace);
 }
 
 PortalNotificationTransport::PortalNotificationTransport(std::chrono::milliseconds closeDelay,
